refact: move swagger docs config to a router class
